requirements.txt
Django~=2.2.4
django-crispy-forms==1.9.0
django-braces==1.14.0
django-model-utils==4.0.0